Removed IGameLoader and GameLoader - these functions are now handled by Game.Setup() Renamed IQuestStage and QuestStage to IQuestStatus and QuestStatus More work implementing framework effects and Game.Run loop

+++ LotR/src/LotR/States/Phases/Any/IPlayersDrawingCards.cs
@@ -10,10 +10,10 @@ namespace LotR.States.Phases.Any
     {
         IEnumerable<Guid> Players { get; }
 
-        byte GetNumberOfCards(Guid playerId);
+        uint GetNumberOfCards(Guid playerId);
         bool PlayerCanDrawCards(Guid playerId);
 
-        void SetNumberOfCards(Guid playerId, byte numberOfCards);
+        void SetNumberOfCards(Guid playerId, uint numberOfCards);
         void EnabledPlayerCardDraw(Guid playerId);
         void DisablePlayerCardDraw(Guid playerId);
     }
